Working product
Most of the bugs have been worked out.

I switched out the WebBrowser control (which is a piece of crap) to a
third party HTML Panel control. Works pretty well.

Stylized the text output with a CSS declaration in the HTML head
section. Sort of annoying and there might be a better way to do this.